test(firestore): reorganize test suite to flat tag-based structure
