+++ packages/google_cloud_firestore/test/vector_test.dart
@@ -12,35 +12,531 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+import 'dart:async';
+
 import 'package:google_cloud_firestore/google_cloud_firestore.dart';
 import 'package:test/test.dart';
 
-import '../fixtures/helpers.dart';
+import 'fixtures/helpers.dart';
 
-/// Integration tests for Vector Search.
-///
-/// These tests require the Firestore emulator to be running.
-/// Start it with: firebase emulators:start --only firestore
 void main() {
-  // Skip all tests if emulator is not configured
-  if (!isFirestoreEmulatorEnabled()) {
-    print(
-      'Skipping Vector integration tests. '
-      'Set FIRESTORE_EMULATOR_HOST environment variable to run these tests.',
+  // Shared Firestore instance for unit tests (no emulator needed)
+  late Firestore firestore;
+
+  setUpAll(() {
+    runZoned(
+      () {
+        firestore = Firestore(
+          settings: const Settings(projectId: 'test-project'),
+        );
+      },
+      zoneValues: {
+        envSymbol: <String, String>{'GOOGLE_CLOUD_PROJECT': 'test-project'},
+      },
     );
-    return;
-  }
+  });
+
+  group('VectorValue', () {
+    test('constructor creates VectorValue from list', () {
+      final vector = VectorValue(const [1.0, 2.0, 3.0]);
+      expect(vector.toArray(), [1.0, 2.0, 3.0]);
+    });
+
+    test('constructor creates immutable copy of list', () {
+      final originalList = [1.0, 2.0, 3.0];
+      final vector = VectorValue(originalList);
+
+      // Modifying original list shouldn't affect VectorValue
+      originalList[0] = 100.0;
+      expect(vector.toArray(), [1.0, 2.0, 3.0]);
+    });
+
+    test('toArray returns a copy', () {
+      final vector = VectorValue(const [1.0, 2.0, 3.0]);
+      final array1 = vector.toArray();
+      final array2 = vector.toArray();
+
+      // Arrays should be equal but not identical
+      expect(array1, array2);
+      expect(identical(array1, array2), false);
+
+      // Modifying returned array shouldn't affect VectorValue
+      array1[0] = 100.0;
+      expect(vector.toArray(), [1.0, 2.0, 3.0]);
+    });
+
+    test('isEqual returns true for equal vectors', () {
+      final vector1 = VectorValue(const [1.0, 2.0, 3.0]);
+      final vector2 = VectorValue(const [1.0, 2.0, 3.0]);
+
+      expect(vector1.isEqual(vector2), true);
+    });
+
+    test('isEqual returns false for different vectors', () {
+      final vector1 = VectorValue(const [1.0, 2.0, 3.0]);
+      final vector2 = VectorValue(const [1.0, 2.0, 4.0]);
+
+      expect(vector1.isEqual(vector2), false);
+    });
+
+    test('isEqual returns false for vectors of different lengths', () {
+      final vector1 = VectorValue(const [1.0, 2.0, 3.0]);
+      final vector2 = VectorValue(const [1.0, 2.0]);
+
+      expect(vector1.isEqual(vector2), false);
+    });
+
+    test('operator == works correctly', () {
+      final vector1 = VectorValue(const [1.0, 2.0, 3.0]);
+      final vector2 = VectorValue(const [1.0, 2.0, 3.0]);
+      final vector3 = VectorValue(const [1.0, 2.0, 4.0]);
+
+      expect(vector1 == vector2, true);
+      expect(vector1 == vector3, false);
+    });
+
+    test('hashCode is consistent for equal vectors', () {
+      final vector1 = VectorValue(const [1.0, 2.0, 3.0]);
+      final vector2 = VectorValue(const [1.0, 2.0, 3.0]);
+
+      expect(vector1.hashCode, vector2.hashCode);
+    });
+
+    test('empty vector is allowed', () {
+      final vector = VectorValue(const []);
+      expect(vector.toArray(), isEmpty);
+    });
+  });
+
+  group('FieldValue.vector', () {
+    test('creates VectorValue', () {
+      final vector = FieldValue.vector([1.0, 2.0, 3.0]);
+
+      expect(vector, isA<VectorValue>());
+      expect(vector.toArray(), [1.0, 2.0, 3.0]);
+    });
+  });
+
+  group('DistanceMeasure', () {
+    test('has correct string values', () {
+      expect(DistanceMeasure.euclidean.value, 'EUCLIDEAN');
+      expect(DistanceMeasure.cosine.value, 'COSINE');
+      expect(DistanceMeasure.dotProduct.value, 'DOT_PRODUCT');
+    });
+  });
+
+  group('VectorQueryOptions', () {
+    test('constructor with required parameters', () {
+      const options = VectorQueryOptions(
+        vectorField: 'embedding',
+        queryVector: [1.0, 2.0, 3.0],
+        limit: 10,
+        distanceMeasure: DistanceMeasure.cosine,
+      );
+
+      expect(options.vectorField, 'embedding');
+      expect(options.queryVector, [1.0, 2.0, 3.0]);
+      expect(options.limit, 10);
+      expect(options.distanceMeasure, DistanceMeasure.cosine);
+      expect(options.distanceResultField, isNull);
+      expect(options.distanceThreshold, isNull);
+    });
+
+    test('constructor with all parameters', () {
+      final options = VectorQueryOptions(
+        vectorField: 'embedding',
+        queryVector: FieldValue.vector([1.0, 2.0, 3.0]),
+        limit: 10,
+        distanceMeasure: DistanceMeasure.euclidean,
+        distanceResultField: 'distance',
+        distanceThreshold: 0.5,
+      );
+
+      expect(options.vectorField, 'embedding');
+      expect(options.queryVector, isA<VectorValue>());
+      expect(options.limit, 10);
+      expect(options.distanceMeasure, DistanceMeasure.euclidean);
+      expect(options.distanceResultField, 'distance');
+      expect(options.distanceThreshold, 0.5);
+    });
+
+    test('equality', () {
+      const options1 = VectorQueryOptions(
+        vectorField: 'embedding',
+        queryVector: [1.0, 2.0, 3.0],
+        limit: 10,
+        distanceMeasure: DistanceMeasure.cosine,
+      );
+
+      const options2 = VectorQueryOptions(
+        vectorField: 'embedding',
+        queryVector: [1.0, 2.0, 3.0],
+        limit: 10,
+        distanceMeasure: DistanceMeasure.cosine,
+      );
+
+      const options3 = VectorQueryOptions(
+        vectorField: 'embedding',
+        queryVector: [1.0, 2.0, 3.0],
+        limit: 5, // different limit
+        distanceMeasure: DistanceMeasure.cosine,
+      );
+
+      expect(options1 == options2, true);
+      expect(options1 == options3, false);
+    });
+  });
+
+  group('Query.findNearest', () {
+    test('validates empty queryVector throws error', () {
+      final query = firestore.collection('collectionId');
+
+      expect(
+        () => query.findNearest(
+          vectorField: 'embedding',
+          queryVector: <double>[],
+          limit: 10,
+          distanceMeasure: DistanceMeasure.euclidean,
+        ),
+        throwsA(isA<ArgumentError>()),
+      );
+    });
+
+    test('validates limit must be positive', () {
+      final query = firestore.collection('collectionId');
+
+      expect(
+        () => query.findNearest(
+          vectorField: 'embedding',
+          queryVector: [10.0, 1000.0],
+          limit: 0,
+          distanceMeasure: DistanceMeasure.euclidean,
+        ),
+        throwsA(isA<ArgumentError>()),
+      );
+
+      expect(
+        () => query.findNearest(
+          vectorField: 'embedding',
+          queryVector: [10.0, 1000.0],
+          limit: -1,
+          distanceMeasure: DistanceMeasure.euclidean,
+        ),
+        throwsA(isA<ArgumentError>()),
+      );
+    });
+
+    test('validates limit must be at most 1000', () {
+      final query = firestore.collection('collectionId');
+
+      expect(
+        () => query.findNearest(
+          vectorField: 'embedding',
+          queryVector: [10.0, 1000.0],
+          limit: 1001,
+          distanceMeasure: DistanceMeasure.euclidean,
+        ),
+        throwsA(isA<ArgumentError>()),
+      );
+    });
+
+    test('accepts VectorValue as queryVector', () {
+      final query = firestore.collection('collectionId');
+      final vectorQuery = query.findNearest(
+        vectorField: 'embedding',
+        queryVector: FieldValue.vector([1.0, 2.0, 3.0]),
+        limit: 10,
+        distanceMeasure: DistanceMeasure.cosine,
+      );
+
+      expect(vectorQuery, isA<VectorQuery<DocumentData>>());
+    });
+
+    test('accepts List<double> as queryVector', () {
+      final query = firestore.collection('collectionId');
+      final vectorQuery = query.findNearest(
+        vectorField: 'embedding',
+        queryVector: [1.0, 2.0, 3.0],
+        limit: 10,
+        distanceMeasure: DistanceMeasure.cosine,
+      );
+
+      expect(vectorQuery, isA<VectorQuery<DocumentData>>());
+    });
+
+    test('accepts FieldPath as vectorField', () {
+      final query = firestore.collection('collectionId');
+      final vectorQuery = query.findNearest(
+        vectorField: FieldPath(const ['nested', 'embedding']),
+        queryVector: [1.0, 2.0, 3.0],
+        limit: 10,
+        distanceMeasure: DistanceMeasure.cosine,
+      );
+
+      expect(vectorQuery, isA<VectorQuery<DocumentData>>());
+    });
+  });
+
+  group('VectorQuery.isEqual', () {
+    test('returns true for equal vector queries', () {
+      final queryA = firestore
+          .collection('collectionId')
+          .where('foo', WhereFilter.equal, 42);
+      final queryB = firestore
+          .collection('collectionId')
+          .where('foo', WhereFilter.equal, 42);
+
+      final vectorQueryA = queryA.findNearest(
+        vectorField: 'embedding',
+        queryVector: [40.0, 41.0, 42.0],
+        limit: 10,
+        distanceMeasure: DistanceMeasure.cosine,
+      );
+
+      final vectorQueryB = queryB.findNearest(
+        vectorField: 'embedding',
+        queryVector: [40.0, 41.0, 42.0],
+        limit: 10,
+        distanceMeasure: DistanceMeasure.cosine,
+      );
+
+      expect(vectorQueryA.isEqual(vectorQueryB), true);
+      expect(vectorQueryA == vectorQueryB, true);
+    });
+
+    test('returns false for different base queries', () {
+      final queryA = firestore
+          .collection('collectionId')
+          .where('foo', WhereFilter.equal, 42);
+      final queryB = firestore.collection('collectionId'); // No where clause
+
+      final vectorQueryA = queryA.findNearest(
+        vectorField: 'embedding',
+        queryVector: [40.0, 41.0, 42.0],
+        limit: 10,
+        distanceMeasure: DistanceMeasure.cosine,
+      );
+
+      final vectorQueryB = queryB.findNearest(
+        vectorField: 'embedding',
+        queryVector: [40.0, 41.0, 42.0],
+        limit: 10,
+        distanceMeasure: DistanceMeasure.cosine,
+      );
+
+      expect(vectorQueryA.isEqual(vectorQueryB), false);
+    });
+
+    test('returns false for different queryVector', () {
+      final queryA = firestore.collection('collectionId');
+      final queryB = firestore.collection('collectionId');
+
+      final vectorQueryA = queryA.findNearest(
+        vectorField: 'embedding',
+        queryVector: [40.0, 41.0, 42.0],
+        limit: 10,
+        distanceMeasure: DistanceMeasure.cosine,
+      );
+
+      final vectorQueryB = queryB.findNearest(
+        vectorField: 'embedding',
+        queryVector: [40.0, 42.0], // Different vector
+        limit: 10,
+        distanceMeasure: DistanceMeasure.cosine,
+      );
+
+      expect(vectorQueryA.isEqual(vectorQueryB), false);
+    });
+
+    test('returns false for different limit', () {
+      final queryA = firestore.collection('collectionId');
+      final queryB = firestore.collection('collectionId');
+
+      final vectorQueryA = queryA.findNearest(
+        vectorField: 'embedding',
+        queryVector: [40.0, 41.0, 42.0],
+        limit: 10,
+        distanceMeasure: DistanceMeasure.cosine,
+      );
+
+      final vectorQueryB = queryB.findNearest(
+        vectorField: 'embedding',
+        queryVector: [40.0, 41.0, 42.0],
+        limit: 1000, // Different limit
+        distanceMeasure: DistanceMeasure.cosine,
+      );
+
+      expect(vectorQueryA.isEqual(vectorQueryB), false);
+    });
+
+    test('returns false for different distanceMeasure', () {
+      final queryA = firestore.collection('collectionId');
+      final queryB = firestore.collection('collectionId');
+
+      final vectorQueryA = queryA.findNearest(
+        vectorField: 'embedding',
+        queryVector: [40.0, 41.0, 42.0],
+        limit: 10,
+        distanceMeasure: DistanceMeasure.cosine,
+      );
+
+      final vectorQueryB = queryB.findNearest(
+        vectorField: 'embedding',
+        queryVector: [40.0, 41.0, 42.0],
+        limit: 10,
+        distanceMeasure: DistanceMeasure.euclidean, // Different measure
+      );
+
+      expect(vectorQueryA.isEqual(vectorQueryB), false);
+    });
+
+    test('returns false for different distanceThreshold', () {
+      final queryA = firestore.collection('collectionId');
+      final queryB = firestore.collection('collectionId');
+
+      final vectorQueryA = queryA.findNearest(
+        vectorField: 'embedding',
+        queryVector: [40.0, 41.0, 42.0],
+        limit: 10,
+        distanceMeasure: DistanceMeasure.euclidean,
+        distanceThreshold: 1.125,
+      );
+
+      final vectorQueryB = queryB.findNearest(
+        vectorField: 'embedding',
+        queryVector: [40.0, 41.0, 42.0],
+        limit: 10,
+        distanceMeasure: DistanceMeasure.euclidean,
+        distanceThreshold: 0.125, // Different threshold
+      );
+
+      expect(vectorQueryA.isEqual(vectorQueryB), false);
+    });
+
+    test('returns false when one has distanceThreshold and other does not', () {
+      final queryA = firestore.collection('collectionId');
+      final queryB = firestore.collection('collectionId');
+
+      final vectorQueryA = queryA.findNearest(
+        vectorField: 'embedding',
+        queryVector: [40.0, 41.0, 42.0],
+        limit: 10,
+        distanceMeasure: DistanceMeasure.euclidean,
+        distanceThreshold: 1,
+      );
+
+      final vectorQueryB = queryB.findNearest(
+        vectorField: 'embedding',
+        queryVector: [40.0, 41.0, 42.0],
+        limit: 10,
+        distanceMeasure: DistanceMeasure.euclidean,
+        // No distanceThreshold
+      );
+
+      expect(vectorQueryA.isEqual(vectorQueryB), false);
+    });
+
+    test('returns false for different distanceResultField', () {
+      final queryA = firestore.collection('collectionId');
+      final queryB = firestore.collection('collectionId');
+
+      final vectorQueryA = queryA.findNearest(
+        vectorField: 'embedding',
+        queryVector: [40.0, 41.0, 42.0],
+        limit: 10,
+        distanceMeasure: DistanceMeasure.euclidean,
+        distanceResultField: 'distance',
+      );
+
+      final vectorQueryB = queryB.findNearest(
+        vectorField: 'embedding',
+        queryVector: [40.0, 41.0, 42.0],
+        limit: 10,
+        distanceMeasure: DistanceMeasure.euclidean,
+        distanceResultField: 'result', // Different field
+      );
+
+      expect(vectorQueryA.isEqual(vectorQueryB), false);
+    });
+
+    test('returns true with distanceResultField as String vs FieldPath', () {
+      final queryA = firestore.collection('collectionId');
+      final queryB = firestore.collection('collectionId');
+
+      final vectorQueryA = queryA.findNearest(
+        vectorField: 'embedding',
+        queryVector: [40.0, 41.0, 42.0],
+        limit: 10,
+        distanceMeasure: DistanceMeasure.euclidean,
+        distanceResultField: 'distance',
+      );
+
+      final vectorQueryB = queryB.findNearest(
+        vectorField: 'embedding',
+        queryVector: [40.0, 41.0, 42.0],
+        limit: 10,
+        distanceMeasure: DistanceMeasure.euclidean,
+        distanceResultField: FieldPath(const ['distance']),
+      );
+
+      expect(vectorQueryA.isEqual(vectorQueryB), true);
+    });
+
+    test('returns true for all distance measures', () {
+      for (final measure in DistanceMeasure.values) {
+        final queryA = firestore.collection('collectionId');
+        final queryB = firestore.collection('collectionId');
+
+        final vectorQueryA = queryA.findNearest(
+          vectorField: 'embedding',
+          queryVector: [1.0],
+          limit: 2,
+          distanceMeasure: measure,
+        );
+
+        final vectorQueryB = queryB.findNearest(
+          vectorField: 'embedding',
+          queryVector: [1.0],
+          limit: 2,
+          distanceMeasure: measure,
+        );
+
+        expect(
+          vectorQueryA.isEqual(vectorQueryB),
+          true,
+          reason: 'Failed for $measure',
+        );
+      }
+    });
+  });
+
+  group('VectorQuery.query', () {
+    test('returns the underlying query', () {
+      final baseQuery = firestore
+          .collection('collectionId')
+          .where('foo', WhereFilter.equal, 42);
+
+      final vectorQuery = baseQuery.findNearest(
+        vectorField: 'embedding',
+        queryVector: [1.0, 2.0, 3.0],
+        limit: 10,
+        distanceMeasure: DistanceMeasure.cosine,
+      );
+
+      expect(vectorQuery.query, baseQuery);
+    });
+  });
 
   group('Vector Integration Tests', () {
-    late Firestore firestore;
+    late Firestore firestoreEmulator;
 
     setUp(() async {
-      firestore = await createFirestore();
+      firestoreEmulator = await createFirestore();
     });
 
     group('write and read vector embeddings', () {
       test('can create document with vector field', () async {
-        final ref = firestore.collection('vector-test').doc();
+        final ref = firestoreEmulator.collection('vector-test').doc();
         await ref.create({
           'vector0': FieldValue.vector([0.0]),
           'vector1': FieldValue.vector([1.0, 2.0, 3.99]),
@@ -58,7 +554,7 @@ void main() {
       });
 
       test('can set document with vector field', () async {
-        final ref = firestore.collection('vector-test').doc();
+        final ref = firestoreEmulator.collection('vector-test').doc();
         await ref.set({
           'vector0': FieldValue.vector([0.0]),
           'vector1': FieldValue.vector([1.0, 2.0, 3.99]),
@@ -81,7 +577,7 @@ void main() {
       });
 
       test('can update document with vector field', () async {
-        final ref = firestore.collection('vector-test').doc();
+        final ref = firestoreEmulator.collection('vector-test').doc();
         await ref.set({'name': 'test'});
         await ref.update({
           'vector3': FieldValue.vector([-1.0, -200.0, -999.0]),
@@ -96,7 +592,7 @@ void main() {
       });
 
       test('VectorValue.isEqual works with retrieved vectors', () async {
-        final ref = firestore.collection('vector-test').doc();
+        final ref = firestoreEmulator.collection('vector-test').doc();
         await ref.set({
           'embedding': FieldValue.vector([1.0, 2.0, 3.0]),
         });
@@ -114,7 +610,7 @@ void main() {
 
       setUp(() async {
         // Create test collection with vector embeddings
-        collection = firestore.collection(
+        collection = firestoreEmulator.collection(
           'vector-search-test-${DateTime.now().millisecondsSinceEpoch}',
         );
 
@@ -309,7 +805,7 @@ void main() {
       });
 
       test('findNearest works with converters', () async {
-        final testCollection = firestore.collection(
+        final testCollection = firestoreEmulator.collection(
           'converter-test-${DateTime.now().millisecondsSinceEpoch}',
         );
 
@@ -339,7 +835,7 @@ void main() {
       });
 
       test('supports findNearest skipping fields of wrong types', () async {
-        final testCollection = firestore.collection(
+        final testCollection = firestoreEmulator.collection(
           'wrong-types-test-${DateTime.now().millisecondsSinceEpoch}',
         );
 
@@ -399,7 +895,7 @@ void main() {
       });
 
       test('findNearest ignores mismatching dimensions', () async {
-        final testCollection = firestore.collection(
+        final testCollection = firestoreEmulator.collection(
           'dimension-test-${DateTime.now().millisecondsSinceEpoch}',
         );
 
@@ -447,7 +943,7 @@ void main() {
       });
 
       test('supports findNearest on non-existent field', () async {
-        final testCollection = firestore.collection(
+        final testCollection = firestoreEmulator.collection(
           'nonexistent-test-${DateTime.now().millisecondsSinceEpoch}',
         );
 
@@ -475,7 +971,7 @@ void main() {
       });
 
       test('supports findNearest with select to exclude vector data', () async {
-        final testCollection = firestore.collection(
+        final testCollection = firestoreEmulator.collection(
           'select-test-${DateTime.now().millisecondsSinceEpoch}',
         );
 
@@ -530,7 +1026,7 @@ void main() {
       });
 
       test('supports findNearest with large dimension vectors', () async {
-        final testCollection = firestore.collection(
+        final testCollection = firestoreEmulator.collection(
           'large-dim-test-${DateTime.now().millisecondsSinceEpoch}',
         );
 
@@ -562,7 +1058,7 @@ void main() {
       });
 
       test('SDK orders vector field same way as backend', () async {
-        final testCollection = firestore.collection(
+        final testCollection = firestoreEmulator.collection(
           'ordering-test-${DateTime.now().millisecondsSinceEpoch}',
         );
 
@@ -617,5 +1113,65 @@ void main() {
         }
       });
     });
-  });
+  }, tags: 'firebase-emulator');
+
+  group('Vector Production Tests', () {
+    late Firestore firestoreProd;
+
+    setUp(() async {
+      firestoreProd = Firestore(
+        settings: const Settings(projectId: 'dart-firebase-admin'),
+      );
+    });
+
+    group('vector search with nested fields', () {
+      test('supports findNearest on vector nested in a map', () async {
+        await runZoned(() async {
+          // Use fixed collection name for production (requires pre-configured index)
+          final collection =
+              firestoreProd.collection('nested-vector-test-prod');
+          final testId = 'test-${DateTime.now().millisecondsSinceEpoch}';
+
+          try {
+            await Future.wait([
+              collection.add({
+                'testId': testId,
+                'nested': {
+                  'embedding': FieldValue.vector([1.0, 1.0]),
+                },
+              }),
+              collection.add({
+                'testId': testId,
+                'nested': {
+                  'embedding': FieldValue.vector([10.0, 10.0]),
+                },
+              }),
+            ]);
+
+            // Query with testId filter for test isolation
+            final vectorQuery = collection
+                .where('testId', WhereFilter.equal, testId)
+                .findNearest(
+                  vectorField: 'nested.embedding',
+                  queryVector: [1.0, 1.0],
+                  limit: 2,
+                  distanceMeasure: DistanceMeasure.euclidean,
+                );
+
+            final res = await vectorQuery.get();
+            expect(res.size, 2);
+          } finally {
+            // Clean up: delete test documents
+            final docs = await firestoreProd
+                .collection('nested-vector-test-prod')
+                .where('testId', WhereFilter.equal, testId)
+                .get();
+            for (final doc in docs.docs) {
+              await doc.ref.delete();
+            }
+          }
+        }, zoneValues: {envSymbol: <String, String>{}});
+      });
+    });
+  }, tags: 'prod');
 }
